告警清理

+++ services/packagegroup/src/bundle_active_group_controller.cpp
@@ -307,10 +307,10 @@ void BundleActiveGroupController::SendCheckBundleMsg(const BundleActiveEvent& ev
 int64_t BundleActiveGroupController::GetMsgKey(const BundleActiveEvent& event, const int32_t& userId)
 {
     std::hash<std::string> hasher;
-    uint32_t bundleNameHash = hasher(event.bundleName_);
-    uint64_t msgKeyHighBit = (uint64_t)bundleNameHash << MSG_KEY_HIGH_BIT;
+    uint64_t bundleNameHash = hasher(event.bundleName_);
+    uint64_t msgKeyHighBit = bundleNameHash << MSG_KEY_HIGH_BIT;
     std::string msgLowHashStr = std::to_string(userId) + std::to_string(bundleNameHash) + std::to_string(event.uid_);
-    uint32_t msgKeyLowBit = hasher(msgLowHashStr);
+    uint64_t msgKeyLowBit = hasher(msgLowHashStr);
     uint64_t msgKey = msgKeyHighBit | msgKeyLowBit;
     return static_cast<int64_t>(msgKey);
 }
